edit DTO fields, edit kafka config

+++ producer-service/src/main/java/org/surro/producerservice/model/UserOrder.java
@@ -1,8 +1,7 @@
 package org.surro.producerservice.model;
 
-import java.math.BigDecimal;
 import java.util.List;
 
-public record UserOrder(int orderId, List<String> dishes, BigDecimal totalPrice) {
+public record UserOrder(int orderId, List<String> dishes, double totalPrice) {
 
 }
